moved files, deleted unnecessary files
