+++ include/ghex/arch_traits.hpp
@@ -13,7 +13,7 @@
 #include <ghex/config.hpp>
 #include <ghex/context.hpp>
 #ifdef GHEX_CUDACC
-#include <ghex/common/cuda_runtime.hpp>
+#include <ghex/device/cuda/runtime.hpp>
 #endif
 
 namespace ghex
