--- include/ghex/common/defs.hpp
@@ -1,18 +1,0 @@
-/*
- * GridTools
- *
- * Copyright (c) 2014-2020, ETH Zurich
- * All rights reserved.
- *
- * Please, refer to the LICENSE file in the root directory.
- * SPDX-License-Identifier: BSD-3-Clause
- *
- */
-#ifndef INCLUDED_GHEX_COMMON_DEFS_HPP
-#define INCLUDED_GHEX_COMMON_DEFS_HPP
-
-#if (defined(__CUDACC__) || defined(__HIP_PLATFORM_HCC__))
-#define GHEX_CUDACC
-#endif
-
-#endif /* INCLUDED_GHEX_COMMON_DEFS_HPP */
