+++ include/ghex/device/cuda/stream.hpp
@@ -12,8 +12,8 @@
 
 #include <ghex/config.hpp>
 #include <ghex/device/cuda/error.hpp>
+#include <ghex/device/cuda/runtime.hpp>
 #include <ghex/util/moved_bit.hpp>
-#include <ghex/common/cuda_runtime.hpp>
 #include <memory>
 
 namespace ghex
