+++ include/ghex/device/cuda/unified_memory_allocator.hpp
@@ -12,7 +12,7 @@
 
 #include <ghex/config.hpp>
 #ifdef GHEX_CUDACC
-#include <ghex/common/cuda_runtime.hpp>
+#include <ghex/device/cuda/runtime.hpp>
 #endif
 
 #include <cstdlib>
