+++ include/ghex/rma/handle.hpp
@@ -18,7 +18,7 @@
 #endif
 #ifdef GHEX_CUDACC
 #include <ghex/rma/cuda/handle.hpp>
-#include <ghex/common/cuda_runtime.hpp>
+#include <ghex/device/cuda/runtime.hpp>
 #endif
 #include <memory>
