+++ include/ghex/structured/cubed_sphere/domain_descriptor.hpp
@@ -10,7 +10,7 @@
  */
 #pragma once
 
-#include <ghex/common/coordinate.hpp>
+#include <ghex/util/coordinate.hpp>
 #include <stdexcept>
 #include <cassert>
