+++ include/ghex/unstructured/user_concepts.hpp
@@ -14,7 +14,7 @@
 #include <ghex/arch_traits.hpp>
 #include <ghex/device/cuda/error.hpp>
 #ifdef GHEX_CUDACC
-#include <ghex/common/cuda_runtime.hpp>
+#include <ghex/device/cuda/runtime.hpp>
 #endif
 
 #include <set>
