+++ include/ghex/util/c_managed_struct.hpp
