+++ test/structured/regular/test_regular_domain.cpp
@@ -19,7 +19,7 @@
 #include <ghex/structured/regular/field_descriptor.hpp>
 #include <ghex/communication_object.hpp>
 #ifdef GHEX_CUDACC
-#include <ghex/common/cuda_runtime.hpp>
+#include <ghex/device/cuda/runtime.hpp>
 #include <gridtools/common/cuda_util.hpp>
 #include <gridtools/common/host_device.hpp>
 #endif
